turns consent question into reusable component #18
