+++ web/templates/consent/new.html.eex
@@ -47,35 +47,7 @@
 
 <%= form_for @conn, consent_path(@conn, :create), [as: :consent], fn f -> %>
   <%= for { question, i } <- Enum.with_index @consent_questions do %>
-    <div class="<%= get_background_colour(i) %> pa4">
-      <span class="ph5"><%= i + 1 %>.</span>
-      <div>
-        <div class="dib v-mid w-80 mr2 ph5">
-          <p>
-            <%= question["field_label"] %>
-          </p>
-        </div>
-        <div class="dib v-mid">
-          <%= radio_button f, question_name(question["field_name"], question["required_field"]), "Yes", class: "yes-consent", id: "yes-consent-#{i+1}" %>
-          <%= label f, question_name(question["field_name"], question["required_field"]), "Yes" %>
-          <%= radio_button f, question_name(question["field_name"], question["required_field"]), "No", class: "no-consent", id: "no-consent-#{i+1}" %>
-          <%= label f, question_name(question["field_name"], question["required_field"]), "No" %>
-        </div>
-        <div id="error-consent-<%= i+1 %>" class="pl3 mh5 dn bg-light-pink w-80 br-pill">
-          <div class="v-mid dib br-100 h3 w3 bg-white tc pt1">
-            <span class="f1">!</span>
-          </div>
-          <div class="dib v-mid ph4 pv1">
-            <p>
-              If you do not consent to this option, you will not be able to join the Research Resource
-            </p>
-            <p>
-              If you are unsure, why not give us a call to talk it through: 020 123 456 - or <a href="#">request a call back</a>
-            </p>
-          </div>
-        </div>
-      </div>
-    </div>
+    <%= component "consent_question", question: question, i: i, f: f %>
   <% end %>
   <div class="pa3 ph5">
     <%= component "button_ok", class: "dn", value: "Next", id: "next-active" %>
